# openapi/Infinity.yaml
openapi: 3.0.3
info:
  version: 0.5.3
  title: Infinity API
  description: "A list of the currently provided Interfaces to interact with the infinity, implementing file operations and sending messages"
security:
  - {}
...
servers:
  - url: "http://{apiRoot}:{port}/v1"
    variables:
      apiRoot:
        default: "localhost"
        description: Base address of the local voyager node main API
      port:
        default: 1633
        description: Service port provided in voyager node config
  - url: "http://{apiRoot}:{port}"
    variables:
      apiRoot:
        default: "localhost"
        description: Base address of the local voyager node main API
      port:
        default: 1633
        description: Service port provided in voyager node config
paths:
...
  "/feeds/{owner}/{topic}":
    post:
      summary: Create an initial feed root manifest
      tags:
        - Feed
      parameters:
        - in: path
          name: owner
          schema:
            $ref: "InfinityCommon.yaml#/components/schemas/EthereumAddress"
          required: true
          description: Owner
        - in: path
          name: topic
          schema:
            $ref: "InfinityCommon.yaml#/components/schemas/HexString"
          required: true
          description: Topic
        - in: query
          name: type
          schema:
            $ref: "InfinityCommon.yaml#/components/schemas/FeedType"
          required: false
          description: "Feed indexing scheme (default: sequence)"
        - $ref: "InfinityCommon.yaml#/components/parameters/InfinityPinParameter"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "InfinityCommon.yaml#/components/schemas/ReferenceResponse"
        "400":
          $ref: "InfinityCommon.yaml#/components/responses/400"
        "401":
          $ref: "InfinityCommon.yaml#/components/responses/401"
        "500":
          $ref: "InfinityCommon.yaml#/components/responses/500"
        default:
          description: Default response
    get:
      summary: Find feed update
      tags:
        - Feed
      parameters:
        - in: path
          name: owner
          schema:
            $ref: "InfinityCommon.yaml#/components/schemas/EthereumAddress"
          required: true
          description: Owner
        - in: path
          name: topic
          schema:
            $ref: "InfinityCommon.yaml#/components/schemas/HexString"
          required: true
          description: Topic
        - in: query
          name: at
          schema:
            type: integer
          required: false
          description: "Timestamp of the update (default: now)"
        - in: query
          name: type
          schema:
            $ref: "InfinityCommon.yaml#/components/schemas/FeedType"
          required: false
          description: "Feed indexing scheme (default: sequence)"
      responses:
        "200":
          description: Latest feed update
          headers:
            "infinity-feed-index":
              $ref: "InfinityCommon.yaml#/components/headers/InfinityFeedIndex"
            "infinity-feed-index-next":
              $ref: "InfinityCommon.yaml#/components/headers/InfinityFeedIndexNext"
          content:
            application/json:
              schema:
                $ref: "InfinityCommon.yaml#/components/schemas/ReferenceResponse"
        "400":
          $ref: "InfinityCommon.yaml#/components/responses/400"
        "401":
          $ref: "InfinityCommon.yaml#/components/responses/401"
        "500":
          $ref: "InfinityCommon.yaml#/components/responses/500"
        default:
          description: Default response